% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDimPrideGdx.R
\name{addDimPrideGdx}
\alias{addDimPrideGdx}
\title{addDimPrideGdx}
\usage{
addDimPrideGdx(file)
}
\arguments{
\item{file}{File name of the gdx file}
}
\description{
Function to expand Pride Model GDX's.
}
\examples{

\dontrun{addDimPrideGdx("bla.gdx")}

}
\author{
Anastasis Giannousakis
}
