% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx-package.R
\docType{package}
\name{gdx-package}
\alias{gdx}
\alias{gdx-package}
\title{gdx: Interface package for GDX files in R}
\description{
A wrapper package for the gdxrrw extending its functionality and allowing to read and write GDX files directly in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/gdx}
  \item \doi{10.5281/zenodo.1158598}
  \item Report bugs at \url{https://github.com/pik-piam/gdx/issues}
}

}
\author{
\strong{Maintainer}: Jan Philipp Dietrich \email{dietrich@pik-potsdam.de}

Authors:
\itemize{
  \item Anastasis Giannousakis \email{giannou@pik-potsdam.de}
  \item Markus Bonsch Bonsch
}

Other contributors:
\itemize{
  \item Lavinia Baumstark Baumstark \email{baumstark@pik-potsdam.de} [contributor]
}

}
