% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdx_rename.R
\name{gdx_rename}
\alias{gdx_rename}
\title{gdx_rename}
\usage{
gdx_rename(file, ..., set_name = NULL)
}
\arguments{
\item{file}{File name of the gdx file in which the objects or set entries should be renamed}

\item{...}{For level=="objects": listing of renamings/deletions that should be done in the form
oldname="newname" (or oldname=0 for delete). Alternatively a named vector could be provided c(oldname="newname",bla="blub",deleteme=0).
For level=="set_names": A named vector should be provided c(newname1="oldname1",newname2="oldname2",newname2="oldname1"). 
It can select only some of the old set entries or use some old entries for more than one new set entry.}

\item{set_name}{If you want to rename entries of a set you have to specify the name of the set. In the default case (set_name=NULL) you can rename objects.}
}
\description{
Function to replace or delete names of variables, parameters or equations in
gdx files (aliases and sets are currently not supported, but support could
be added if required).
}
\examples{

\dontrun{

#list all objects in the given gdx file
gdx_rename("bla.gdx")

#replace oldname with newname and oldname2 with newname2
gdx_rename("bla.gdx",oldname="newname",oldname2="newname2")

#delete "deleteme"
gdx_rename("bla.gdx",deleteme=0)

# rename and select set entries of the set "testset"
gdx_rename("bla.gdx",set_name="testset",c(newentry1="oldentry1",newentry2="oldentry2"))
}

}
\seealso{
\code{\link{readGDX}}, \code{\link{writeGDX}}
}
\author{
Jan Philipp Dietrich, Lavinia Baumstark
}
