% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inp.R
\name{inp}
\alias{inp}
\title{inp - \bold{DEPRECEATED!}}
\usage{
inp(gdx, name, ..., react = "warning", as.magpie = FALSE)
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx file
(file name is recommended as this speeds up the code)}

\item{name}{name of the parameter or set that should be read}

\item{...}{Additional names. If the attempt to read 'name' fails, these will
be tried in the order given by the user until one is successfully read.}

\item{react}{determines the reaction, when the parameter or set does not
exist. Available options are "warning" (NULL is returned and a warning is
send that the parameter or set is missing), "silent" (NULL is returned, but
no warning is given) and "error" (The function throws out an error)}

\item{as.magpie}{If TRUE the content is returned as a MAgPIE object,
otherwise as array}
}
\value{
The parameter if it exists, otherwise NULL
}
\description{
This function is depreceated, please do not use it! Please see the note
below for more information. Function to savely read parameters and sets from
GDX file. The function creates a warning if the parameter does not exist in
the gdx file. Please use this function when you write own GDX output
functions.
}
\note{
\bold{This function is depreceated! Please use
\code{\link{readGDX}(...,format="first_found")} instead! This will basically
give you the same functionality with the only difference that it will return
a magpie object instead of an array.}
}
\examples{

\dontrun{inp("fulldata.gdx","x_modelstat")}

}
\seealso{
\code{\link{out}}, \code{\link{readGDX}}
}
\author{
Jan Philipp Dietrich
}
