% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals_check.R
\name{marginals_check}
\alias{marginals_check}
\title{marginals_check}
\usage{
marginals_check(gdx, ifilter=NULL, efilter=NULL, limit=1e5,
scientific=TRUE, unlist=FALSE, name="o(v|q)*")
}
\arguments{
\item{gdx}{a GDX list as created by readGDX, or the file name of a gdx file}

\item{ifilter}{Inclusion filter. A vector of strings containing regular
expressions which describe strings that have to be part of the found
marginal names. Otherwise they will be excluded.}

\item{efilter}{Exclusion filter. A vector of strings containing regular
expressions which describe strings must not be part of the found marginal
names. Otherwise they will be excluded.}

\item{limit}{Lower limit for the absolute value of the marginal in order to
be part of the output.}

\item{scientific}{Boolean which decides whether the marginals should be
written scientfically (e.g. 1e+2) or the default output scheme should be
used.}

\item{unlist}{If TRUE a vector sorted by the rank of the marginal will be
returned, otherwise a structured list of outputs will be returned.}

\item{name}{search string defining the objects that should be read from gdx
file, with *-autocompletion. Can also be a vector containing more than one
search strings}
}
\value{
A vector or list containing all marginals which absolute values are
above the given limit and which agree with the given filters.
}
\description{
This function returns a list of the highest marginals found in gdx file
allowing to detect the most critical constraints in a given solution.
}
\examples{

\dontrun{marginals_check("fulldata.gdx")}

}
\seealso{
\code{\link{readGDX}}
}
\author{
Jan Philipp Dietrich
}
