% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGDX.R
\name{writeGDX}
\alias{writeGDX}
\title{writeGDX}
\usage{
writeGDX(x,file,period_with_y = TRUE)
}
\arguments{
\item{x}{A list of objects or a single object that should be written to a
gdx file. The format is identical to what you get as output by using
\code{\link{readGDX}} (all formats provided by readGDX are supported by
writeGDX as well).}

\item{file}{File name of the gdx file}

\item{period_with_y}{Keep "y" in period dimension}
}
\description{
Function to write gdx files in R.
}
\examples{

\dontrun{writeGDX(readGDX("input.gdx"),"test.gdx")}



}
\seealso{
\code{\link{readGDX}}
}
\author{
Jan Philipp Dietrich
}
